input: handle errormessage property
If the errorMessage is set via HTML instead of via `reRender()`
or a function call then the DOM will lower case the attribute
name completely so the actual attribute name is `errormessage` and
not `errorMessage`.

With this logic we respect the JS property first and we respect
the HTML attribute second. Because of that behavior we have to
remove the default into the `render()` method itself, otherwise
the default overwrites the DOM attribute.