checkbox: simplify checkbox
This updates the logic in the getter and forces all other
code to use the getter & setter when reading the current
value or trying to set the current value.

This fixes a bug where `this.props.checked === true` and
`this.state.checked === false` and we attempt to toggle
the checkbox we toggle it from false => true but render it
from true => true which causes visual confusion.