input: Do not relay `input` event.
On the `TonicInput` component we trigger the input event twice.

For focus & blur the `relay()` function is necessary since the
DOM `blur` & `focus` events do not bubble.

However the `input` event already bubbles so we do not have to
relay it, otherwise the outer component event listener gets
called twice and does duplicate work.