windowed: Add prefetchBottom() support.
If the user implements `prefetchBottom()` it will get called
if the user should prefetch.

If the user implements `renderLoadingBottom()` it will be called
and appended to the bottom of the windowed component to show
a loading indicator.

There is a new method called `pushEOL()` that the user can call
to tell Windowed that there is nothing left to prefetch on
the bottom of the component and everything has been fetched.