windowed: Add prefetchTop & maxRowsLength
This adds the notion of a max number of rows in the Windowed
component.

If `prefetchTop` is not implemented then the HEAD of the array
will be truncated and you will not be able to scroll back
further then the LAST MAX_ROWS_LENGTH rows that have been
rendered in the Windowed component.

If `prefetchTop` is implemented then we will call your method
and you are responsible for prepending data to the array.

When we truncate we respect the most recent scroll direction and
will truncate the array at either the TOP or BOTTOM based on
whether we are scrolling at the BOTTOM or scrolling at the TOP.

There is a potential race condition with the truncation if the
user very quickly jumps from scrolling top to bottom or bottom to
top.

YOU MUST read `this.rows[0]` or `this.rows[this.rows.length - 1]`
in your `prefetchTop()` or `prefetchBottom()` to determine the
offset at with to fetch data.