windowed: When truncating, move pages.
When we change the scrollTop we actually change the current
pageIndex by a large jump, this is because we have truncated
the array.

We move the page objects to the new index they should be in.

This will out the next render to get existing pages and update
them instead of creating new pages and rendering them with
innerHTML.