Tabs: Call onTabShow if exists
This allows a Tonic component embedded inside a Tab to
be notified when it is made visible.

This could be implemented manually but since the Tab is toggled
both by clicks and by keyboard presses the outer component
would have to either handle both or have a buggy keyboard impl.