chart: Support bundling of tonic-chart
The tonic-chart component currently uses dynamicRequire to
load the chart.js library.

This confuses the bundling tooling like `noderify` which means
that the `tonic-chart` component works in our electron app EXCEPT
when we bundle it into a package app binary with an `app.asar`.

By passing the `require('chart.js')` library as either a HTML
attribute or as a `setChart()` method we can require the `chart.js`
library in our app, have that picked up by the bundler without
a dynamic require statement and have the `tonic-chart` working
in the packaged version of our app.