src: Support renderAll flag.
This is useful if you have two tabs that are connected.

For example you have a textual editor and a visual editor
in two tabs and you want their changes to be applied to each
other.

In this case you always want to render all elements in the DOM
but mark them as hidden or not hidden with CSS for visual
purposes.